Source Maintenance: Move debugs()-related code into src/debug/
The new library is needed to facilitate debugs() reuse and, perhaps more
importantly, reuse of other Squid code that already uses debugs() -- all
without build-breaking hacks of copying debug.cc source files and very
dangerous stub use in production running code.

The Raw class and asHex() I/O manipulator are not specific to debugs()
streams. They are now in src/base/.

Also removed manual setting of DIST_SUBDIRS in src/Makefile.am. Automake
documentation promises that "If SUBDIRS is defined conditionally using
Automake conditionals, Automake will define DIST_SUBDIRS automatically
from the possible values of SUBDIRS in all conditions." The tests have
confirmed that Automake does exactly that. We could remove some other
DIST_SUBDIRS, but that seems like an out of scope task. This particular
DIST_SUBDIRS had to be modified anyway...

Also removed test-suite/debug, addressing an XXX. That test was
misplaced and not really doing anything useful -- the compiler already
tests what test-suite/debug was testing (and much more).

TODO: Adjust the moved code to use Debug namespace. Those adjustments
are likely to trigger debug/Stream.h reorganization, simplifying exposed
debugging API. They deserve a dedicated PR.

TODO: Convert more helpers to use the new debugging library. This change
only adjusts pinger because that adjustment did not require other
significant changes.

XXX: Missing Debug.h => debug/Stream.h and similar "boring" #include
changes.
