+++ src/Makefile.am
@@ -31,18 +31,36 @@ LOADABLE_MODULES_SOURCES = \
 	LoadableModules.cc \
 	LoadableModules.h
 
-SUBDIRS = mem base anyp helper dns ftp parser comm error eui acl format clients sbuf servers fs repl store DiskIO proxyp
-DIST_SUBDIRS = mem base anyp helper dns ftp parser comm error eui acl format clients sbuf servers fs repl store DiskIO proxyp
+SUBDIRS = \
+	mem \
+	debug \
+	base \
+	anyp \
+	helper \
+	dns \
+	ftp \
+	parser \
+	comm \
+	error \
+	eui \
+	acl \
+	format \
+	clients \
+	sbuf \
+	servers \
+	fs \
+	repl \
+	store \
+	DiskIO \
+	proxyp
 
 if ENABLE_AUTH
 SUBDIRS += auth
 AUTH_LIBS= auth/libauth.la
 AUTH_ACL_LIBS= auth/libacls.la
 endif
-DIST_SUBDIRS += auth
 
 SUBDIRS	+= http ip icmp ident log ipc mgr
-DIST_SUBDIRS += http ip icmp ident log ipc mgr
 
 SSL_LIBS=
 if ENABLE_SSL
@@ -51,7 +69,6 @@ SSL_LIBS += \
 	ssl/libsslsquid.la \
 	ssl/libsslutil.la
 endif
-DIST_SUBDIRS += ssl security
 SUBDIRS += security
 
 SNMP_ALL_SOURCE = \
@@ -67,12 +84,10 @@ SNMP_LIBS = snmp/libsnmp.la $(SNMPLIB)
 else
 SNMP_SOURCE =
 endif
-DIST_SUBDIRS += snmp
 
 if ENABLE_ADAPTATION
 SUBDIRS += adaptation
 endif
-DIST_SUBDIRS += adaptation
 
 if ENABLE_ESI
 SUBDIRS += esi
@@ -84,7 +99,6 @@ ESI_LIBS = \
 else
 ESI_LIBS =
 endif
-DIST_SUBDIRS += esi
 
 DELAY_POOL_ALL_SOURCE = \
 	BandwidthBucket.cc \
@@ -242,8 +256,6 @@ squid_SOURCES = \
 	CpuAffinityMap.h \
 	CpuAffinitySet.cc \
 	CpuAffinitySet.h \
-	Debug.h \
-	DebugMessages.h \
 	Downloader.cc \
 	Downloader.h \
 	ETag.cc \
@@ -372,7 +384,6 @@ squid_SOURCES = \
 	client_side_reply.h \
 	client_side_request.cc \
 	client_side_request.h \
-	debug.cc \
 	defines.h \
 	dlink.cc \
 	dlink.h \
@@ -543,6 +554,7 @@ squid_LDADD = \
 	log/liblog.la \
 	format/libformat.la \
 	sbuf/libsbuf.la \
+	debug/libdebug.la \
 	$(XTRA_OBJS) \
 	$(REPL_OBJS) \
 	$(NETTLELIB) \
@@ -614,7 +626,6 @@ ufsdump_SOURCES = \
 	String.cc \
 	cbdata.cc \
 	cbdata.h \
-	debug.cc \
 	dlink.cc \
 	dlink.h \
 	tests/stub_fd.cc \
@@ -639,6 +650,7 @@ ufsdump_LDADD = \
 	fs/libfs.la \
 	ipc/libipc.la \
 	mgr/libmgr.la \
+	debug/libdebug.la \
 	$(XTRA_OBJS) \
 	$(REPL_OBJS) \
 	$(NETTLELIB) \
@@ -1935,7 +1947,6 @@ tests_test_http_range_SOURCES = \
 	client_side.h \
 	client_side_reply.cc \
 	client_side_request.cc \
-	debug.cc \
 	dlink.cc \
 	dlink.h \
 	errorpage.cc \
@@ -2059,6 +2070,7 @@ tests_test_http_range_LDADD = \
 	base/libbase.la \
 	mgr/libmgr.la \
 	sbuf/libsbuf.la \
+	debug/libdebug.la \
 	store/libstore.la \
 	$(SNMP_LIBS) \
 	$(top_builddir)/lib/libmisccontainers.la \
@@ -2076,7 +2088,6 @@ tests_test_http_range_LDFLAGS = $(LIBADD_DL)
 
 check_PROGRAMS += tests/testHttp1Parser
 tests_testHttp1Parser_SOURCES = \
-	Debug.h \
 	tests/stub_HelperChildConfig.cc \
 	tests/testHttp1Parser.cc \
 	tests/testHttp1Parser.h \
@@ -2329,7 +2340,6 @@ tests_testHttpRequest_SOURCES = \
 	client_side.h \
 	client_side_reply.cc \
 	client_side_request.cc \
-	debug.cc \
 	dlink.cc \
 	dlink.h \
 	errorpage.cc \
@@ -2452,6 +2462,7 @@ tests_testHttpRequest_LDADD = \
 	format/libformat.la \
 	store/libstore.la \
 	sbuf/libsbuf.la \
+	debug/libdebug.la \
 	$(REPL_OBJS) \
 	$(ADAPTATION_LIBS) \
 	$(ESI_LIBS) \
@@ -2636,7 +2647,6 @@ tests_testCacheManager_SOURCES = \
 	client_side.h \
 	client_side_reply.cc \
 	client_side_request.cc \
-	debug.cc \
 	dlink.cc \
 	dlink.h \
 	errorpage.cc \
@@ -2763,6 +2773,7 @@ tests_testCacheManager_LDADD = \
 	mem/libmem.la \
 	store/libstore.la \
 	sbuf/libsbuf.la \
+	debug/libdebug.la \
 	$(top_builddir)/lib/libmisccontainers.la \
 	$(top_builddir)/lib/libmiscencoding.la \
 	$(top_builddir)/lib/libmiscutil.la \
