+++ src/base/Raw.cc
@@ -1,0 +1,62 @@
+/*
+ * Copyright (C) 1996-2022 The Squid Software Foundation and contributors
+ *
+ * Squid software is distributed under GPLv2+ license and includes
+ * contributions from numerous individuals and organizations.
+ * Please see the COPYING and CONTRIBUTORS files for details.
+ */
+
+/* DEBUG: section 00    Debug Routines */
+
+#include "squid.h"
+#include "base/Raw.h"
+#include "debug/Stream.h"
+
+#include <algorithm>
+
+void
+PrintHex(std::ostream &os, const char *data, const size_t n)
+{
+    if (!n)
+        return;
+    assert(data);
+
+    const auto savedFill = os.fill('0');
+    const auto savedFlags = os.flags(); // std::ios_base::fmtflags
+    os << std::hex;
+    std::for_each(data, data + n,
+    [&os](const char &c) { os << std::setw(2) << static_cast<uint8_t>(c); });
+    os.flags(savedFlags);
+    os.fill(savedFill);
+}
+
+std::ostream &
+Raw::print(std::ostream &os) const
+{
+    if (label_)
+        os << ' ' << label_ << '[' << size_ << ']';
+
+    if (!size_)
+        return os;
+
+    // finalize debugging level if no level was set explicitly via minLevel()
+    const int finalLevel = (level >= 0) ? level :
+                           (size_ > 40 ? DBG_DATA : Debug::SectionLevel());
+    if (finalLevel <= Debug::SectionLevel()) {
+        if (label_)
+            os << '=';
+        else if (useGap_)
+            os << ' ';
+        if (data_) {
+            if (useHex_)
+                PrintHex(os, data_, size_);
+            else
+                os.write(data_, size_);
+        } else {
+            os << "[null]";
+        }
+    }
+
+    return os;
+}
+
