+++ src/debug/Makefile.am
@@ -1,0 +1,17 @@
+## Copyright (C) 1996-2022 The Squid Software Foundation and contributors
+##
+## Squid software is distributed under GPLv2+ license and includes
+## contributions from numerous individuals and organizations.
+## Please see the COPYING and CONTRIBUTORS files for details.
+##
+
+include $(top_srcdir)/src/Common.am
+include $(top_srcdir)/src/TestHeaders.am
+
+noinst_LTLIBRARIES = libdebug.la
+
+libdebug_la_SOURCES = \
+	debug.cc \
+	Messages.h \
+	Stream.h
+
