+++ src/debug/debug.cc
@@ -11,8 +11,7 @@
 #include "squid.h"
 #include "base/Optional.h"
 #include "base/TextException.h"
-#include "Debug.h"
-#include "DebugMessages.h"
+#include "debug/Stream.h"
 #include "fd.h"
 #include "ipc/Kids.h"
 #include "SquidTime.h"
@@ -32,7 +31,6 @@ bool Debug::log_syslog = false;
 int Debug::Levels[MAX_DEBUG_SECTIONS];
 char *Debug::cache_log = NULL;
 int Debug::rotateNumber = -1;
-DebugMessages TheDebugMessages;
 
 /// a counter related to the number of debugs() calls
 using DebugRecordCount = uint64_t;
@@ -1360,51 +1358,5 @@ ForceAlert(std::ostream& s)
 {
     Debug::ForceAlert();
     return s;
-}
-
-void
-PrintHex(std::ostream &os, const char *data, const size_t n)
-{
-    if (!n)
-        return;
-    assert(data);
-
-    const auto savedFill = os.fill('0');
-    const auto savedFlags = os.flags(); // std::ios_base::fmtflags
-    os << std::hex;
-    std::for_each(data, data + n,
-    [&os](const char &c) { os << std::setw(2) << static_cast<uint8_t>(c); });
-    os.flags(savedFlags);
-    os.fill(savedFill);
-}
-
-std::ostream &
-Raw::print(std::ostream &os) const
-{
-    if (label_)
-        os << ' ' << label_ << '[' << size_ << ']';
-
-    if (!size_)
-        return os;
-
-    // finalize debugging level if no level was set explicitly via minLevel()
-    const int finalLevel = (level >= 0) ? level :
-                           (size_ > 40 ? DBG_DATA : Debug::SectionLevel());
-    if (finalLevel <= Debug::SectionLevel()) {
-        if (label_)
-            os << '=';
-        else if (useGap_)
-            os << ' ';
-        if (data_) {
-            if (useHex_)
-                PrintHex(os, data_, size_);
-            else
-                os.write(data_, size_);
-        } else {
-            os << "[null]";
-        }
-    }
-
-    return os;
 }
